chore: cleaning events models
