--- application_sdk/events/workflow_events.py
@@ -1,8 +1,0 @@
-"""Workflow event models.
-
-This module contains event models specific to workflow operations.
-Currently a placeholder for future workflow-related events.
-"""
-
-# This file is a placeholder for future workflow-specific event models
-# that may be added as the application grows.
